refactor: CategoryNameVerifier 분리
