+++ presentation/src/main/java/com/pomonyang/mohanyang/presentation/screen/home/category/CategorySettingViewModel.kt
@@ -4,10 +4,8 @@ import androidx.lifecycle.SavedStateHandle
 import androidx.navigation.toRoute
 import com.pomonyang.mohanyang.presentation.base.BaseViewModel
 import com.pomonyang.mohanyang.presentation.screen.home.CategorySetting
-import com.pomonyang.mohanyang.presentation.screen.onboarding.naming.ValidationResult
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
-import timber.log.Timber
 
 @HiltViewModel
 class CategorySettingViewModel @Inject constructor(
@@ -60,19 +58,5 @@ class CategorySettingViewModel @Inject constructor(
                 setEffect(CategorySettingSideEffect.DismissCategoryIconBottomSheet)
             }
         }
-    }
-
-    fun validateCategoryName(name: String): ValidationResult {
-        // TODO 기존 카테고리 이름 리스트 가져와서 중복값 확인하는 로직 추가
-        Timber.d("${name.length} => ${name.length in NAME_MIN_LENGTH..NAME_MAX_LENGTH}")
-        return ValidationResult(
-            isValid = name.length in NAME_MIN_LENGTH..NAME_MAX_LENGTH,
-            message = if (name.length <= NAME_MAX_LENGTH) "" else "최대 ${NAME_MAX_LENGTH}자리까지 입력할 수 있어요.",
-        )
-    }
-
-    companion object {
-        private const val NAME_MAX_LENGTH = 10
-        private const val NAME_MIN_LENGTH = 1
     }
 }
